api-server/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.3
